Adds a Singleton class
This update adds a singleton class so that only a single instance of a device can be created. the benifit to this would be if you connect to a device in on place in your code and you connect to the same device in a different place it does not create a new device instance, instead it returns the already created instance. It makes it easier for doing things like checking if new devices were added to the system. But now also offers the ability to cache specific data sets from the device and know that all stored instances of the device will contain the new cached data set.